requirements.txt
arrow==0.15.8
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.8.0
django-fernet-fields==0.6
djangorestframework==3.13.1
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
numpy==1.23.1
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.21
python-dateutil==2.7.0
python-decouple==3.4
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tensorflow==2.8.0
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0